+++ dominio/src/main/java/br/com/gestao_escola/dominio/entidade/disciplina/Turma.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Data
 public class Turma {
 
-    private float id;
+    private int id;
 
     private String nome;
