+++ dominio/src/main/java/br/com/gestao_escola/dominio/entidade/forum/Forum.java
@@ -9,7 +9,7 @@ import java.util.stream.Collector;
 @Data
 public class Forum<T>{
 
-    private float id;
+    private int id;
 
     private String mensagem;
