+++ persistencia/src/main/java/br/com/gestao_escola/persistencia/repositorio/AulaRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AulaRepositorio extends JpaRepository<AulaEntidade, Float> {
+public interface AulaRepositorio extends JpaRepository<AulaEntidade, Integer> {
 }
