+++ persistencia/src/main/java/br/com/gestao_escola/persistencia/repositorio/OrganizaForumRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OrganizaForumRepositorio extends JpaRepository<OrganizaForumEntidade, Float> {
+public interface OrganizaForumRepositorio extends JpaRepository<OrganizaForumEntidade, Integer> {
 }
